refactor: Move storage/compression to storage/level/local/encoding/compression
