+++ internal/pkg/service/stream/storage/level/local/factory.go
@@ -3,7 +3,7 @@ package local
 import (
 	"path/filepath"
 
-	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/compression"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/level/local/encoding/compression"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/level/local/model"
 )
