+++ internal/pkg/service/stream/storage/level/staging/model/slice.go
@@ -1,7 +1,7 @@
 package model
 
 import (
-	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/compression"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/level/local/encoding/compression"
 )
 
 type Slice struct {
