+++ internal/pkg/service/stream/storage/model/naming.go
@@ -1,7 +1,7 @@
 package model
 
 import (
-	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/compression"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/stream/storage/level/local/encoding/compression"
 	"github.com/keboola/keboola-as-code/internal/pkg/utils/errors"
 )
